Springified without boot.:)

+++ sportsbettingapplicationmaven/sportsbetting-application-app/src/main/java/com/example/sportsbetting/controller/UserController.java
@@ -1,12 +1,20 @@
 package com.example.sportsbetting.controller;
 
 import com.example.sportsbetting.domain.Currency;
+import com.example.sportsbetting.domain.OutcomeOdd;
 import com.example.sportsbetting.domain.Player;
 import com.example.sportsbetting.domain.Wager;
+import com.example.sportsbetting.dto.OutcomeOddDto;
+import com.example.sportsbetting.dto.PlayerDto;
+import com.example.sportsbetting.dto.SportEventDto;
+import com.example.sportsbetting.dto.WagerDto;
 import com.example.sportsbetting.exception.CurrencyMismatchException;
-import com.example.sportsbetting.service.ISportsBettingService;
+import com.example.sportsbetting.model.DetailsRequest;
+import com.example.sportsbetting.model.WagerRequest;
 import com.example.sportsbetting.service.SportsBettingService;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.servlet.ModelAndView;
@@ -16,13 +24,18 @@ import java.math.BigDecimal;
 
 @Controller
 public class UserController {
-    private ISportsBettingService service = SportsBettingService.GetInstance();
+    @Autowired
+    private SportsBettingService service;
 
     @GetMapping("/user")
     public ModelAndView index(HttpSession session) {
         if (session.getAttribute("Id") == null) {
             return new ModelAndView("redirect:/login");
         }
+
+        Integer playerId = (Integer) session.getAttribute("Id");
+        PlayerDto player = service.findPlayerDto(playerId);
+        Iterable<WagerDto> wagers = service.findAllPlayerWagerDtos(playerId);
 
         return new ModelAndView("redirect:/user/dashboard");
     }
@@ -33,7 +46,13 @@ public class UserController {
             return new ModelAndView("redirect:/login");
         }
 
-        return new ModelAndView("user/dashboard");
+        Integer playerId = (Integer) session.getAttribute("Id");
+        PlayerDto player = service.findPlayerDto(playerId);
+        Iterable<WagerDto> wagers = service.findAllPlayerWagerDtos(playerId);
+
+        return new ModelAndView("user/dashboard")
+                .addObject("player", player)
+                .addObject("wagers", wagers);
     }
 
     @GetMapping("/user/wager")
@@ -42,7 +61,9 @@ public class UserController {
             return new ModelAndView("redirect:/login");
         }
 
-        return new ModelAndView("user/wager");
+        Iterable<OutcomeOddDto> odds = service.findAllOutcomeOddDtos();
+        return new ModelAndView("user/wager")
+                .addObject("odds", odds);
     }
 
     @GetMapping("/user/events")
@@ -51,7 +72,8 @@ public class UserController {
             return new ModelAndView("redirect:/login");
         }
 
-        return new ModelAndView("user/events");
+        Iterable<SportEventDto> events = service.findAllSportEventDtos();
+        return new ModelAndView("user/events").addObject("events", events);
     }
 
     @GetMapping("/user/calculate")
@@ -65,7 +87,7 @@ public class UserController {
         } catch (CurrencyMismatchException e) {
 
         }
-        return new ModelAndView("user/dashboard");
+        return new ModelAndView("redirect:/user/dashboard");
     }
 
     @GetMapping("/user/logout")
@@ -78,26 +100,34 @@ public class UserController {
     }
 
     @PostMapping("/user/wager")
-    public ModelAndView addWager(HttpSession session, Integer odd, Currency currency, BigDecimal amount) {
+    public ModelAndView addWager(HttpSession session, WagerRequest wagerRequest, BindingResult bindingResult) {
         if (session.getAttribute("Id") == null) {
             return new ModelAndView("redirect:/login");
         }
 
-        try {
-            Integer playerId = (Integer) session.getAttribute("Id");
-            Player player = service.findPlayer(playerId);
-
-            Wager wager = new Wager();
-            wager.setPlayer(player);
-            wager.setCurrency(player.getCurrency());
-            wager.setOutcomeOdd(service.findOutcomeOdd(odd));
-            wager.setAmount(amount);
-
-            service.saveWager(wager);
-            return new ModelAndView("redirect:/user/dashboard");
-        } catch (Exception e) {
+        if(bindingResult.hasErrors() || !wagerRequest.validate()) {
             return new ModelAndView("redirect:/user/wager");
         }
+
+        try {
+            int playerId = (int)session.getAttribute("Id");
+            Player player = service.findPlayer(playerId);
+            OutcomeOdd outcomeOdd = service.findOutcomeOdd(wagerRequest.getOddId());
+
+            if(outcomeOdd != null) {
+                Wager wager = new Wager();
+                wager.setPlayer(player);
+                wager.setCurrency(player.getCurrency());
+                wager.setOutcomeOdd(outcomeOdd);
+                wager.setAmount(wagerRequest.getAmount());
+
+                service.saveWager(wager);
+                return new ModelAndView("redirect:/user/dashboard");
+            }
+        } catch (Exception e) {
+        }
+
+        return new ModelAndView("redirect:/user/wager");
     }
 
     @GetMapping(value = "/user/wager/delete")
@@ -106,35 +136,37 @@ public class UserController {
             return new ModelAndView("redirect:/login");
         }
 
-        try {
-            service.deleteWager(wagerId);
-        } catch (CurrencyMismatchException c) {
-
+        if (wagerId != null) {
+            try {
+                service.deleteWager(wagerId);
+            } catch (CurrencyMismatchException ignored) {
+            }
         }
 
         return new ModelAndView("redirect:/user/dashboard");
     }
 
     @PostMapping("/user/dashboard")
-    public ModelAndView saveDetails(HttpSession session, Player player) {
+    public ModelAndView saveDetails(HttpSession session, DetailsRequest detailsRequest, BindingResult bindingResult) {
         if (session.getAttribute("Id") == null) {
             return new ModelAndView("redirect:/login");
         }
 
-        try {
-            Integer playerId = (Integer) session.getAttribute("Id");
-            Player dbPlayer = service.findPlayer(playerId);
+        if(!bindingResult.hasErrors() && detailsRequest.validate()) {
+            try {
+                Integer playerId = (Integer) session.getAttribute("Id");
+                Player dbPlayer = service.findPlayer(playerId);
 
-            dbPlayer.setName(player.getName());
-            dbPlayer.setBirth(player.getBirth());
-            dbPlayer.setAccountNumber(player.getAccountNumber());
-            dbPlayer.setCurrency(player.getCurrency());
-            dbPlayer.setBalance(player.getBalance());
+                dbPlayer.setName(detailsRequest.getName());
+                dbPlayer.setBirth(detailsRequest.getBirth());
+                dbPlayer.setAccountNumber(detailsRequest.getAccountNumber());
+                dbPlayer.setCurrency(detailsRequest.getCurrency());
+                dbPlayer.setBalance(detailsRequest.getBalance());
 
-            service.savePlayer(dbPlayer);
-        } catch (Exception e) {
+                service.savePlayer(dbPlayer);
+            } catch (Exception e) {
+            }
         }
-
         return new ModelAndView("redirect:/user/dashboard");
     }
 }
